Split source code into multiple files
